requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
cryptography==3.4.7
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-nested-admin==3.3.3
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
netmiko==3.4.0
ntc-templates==2.0.0
paramiko==2.7.2
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyserial==3.5
python-dotenv==0.17.0
python-monkey-business==1.0.0
pytz==2021.1
scp==0.13.3
six==1.16.0
sqlparse==0.4.1
tenacity==7.0.0
textfsm==1.1.0
waitress==2.0.0
ipdb==0.13.13
